Create inheritance.cpp
This C++ program demonstrates the concept of multiple inheritance through a hybrid calculator that combines the functionalities of both a simple calculator and a scientific calculator. It defines two base classes: simplecalculator and scientificcalc. The simplecalculator class allows the user to perform basic arithmetic operations such as addition, subtraction, multiplication, and division on two integers, a and b. It contains member functions setnum1() to input the values and operation1() to perform and display the results of the operations. The scientificcalc class provides scientific functions such as cosine, sine, exponential, and tangent of a given value a, using the C++ <math.h> library. It also includes setnum2() and operation2() functions for input and operations respectively.

The hybrid class inherits from both simplecalculator and scientificcalc, allowing it to access and use the functionalities of both calculators. In the main() function, individual objects of simplecalculator and scientificcalc are created to demonstrate their functionalities independently. Finally, an object of the hybrid class is used to show that it can perform both types of calculations by invoking both setnum1()/operation1() and setnum2()/operation2(). This program effectively illustrates how multiple inheritance can be used in C++ to combine different features into a single class.