Create write.cpp
This C++ program demonstrates how to perform basic file output operations using the ofstream class from the <fstream> header. The program begins by including the necessary headers <iostream> and <fstream>, which are required for standard input/output and file handling respectively. Inside the main() function, a string variable named st is initialized with the text "hello every one this is file I/O in c++ ". Then, an ofstream (output file stream) object named out is created and associated with a file named sample.txt. Using the output stream operator <<, the content of the string st is written to the file. After writing the data, the out.close() function is called to close the file, ensuring that all data is properly saved and system resources are released. As a result, this program creates (or overwrites) a file named sample.txt with the given text, showcasing how to write data to a file in C++.